# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RTB AI Hub API
  version: 1.0.0
  description: |
    AI-powered automation hub that orchestrates Figma design → Jira tickets → auto development → code review → deployment → monitoring.

    The API consists of two services:
    - **Webhook Listener** (port 4000): Receives webhook events from Figma, Jira, GitHub, and Datadog, validates payloads via Zod schemas, and queues them for async processing via BullMQ.
    - **Auth Service** (port 4001): Handles Google Workspace OAuth login, session management, and encrypted credential storage (AES-256-GCM) for per-user API keys.

    ### Multi-Environment Support
    All webhook endpoints support environment routing via `?env=int|stg|prd` query parameter or `X-Env` header. Events are processed by environment-specific MCP server containers with separate credentials. Default environment is `int`.

    ### Authentication Flow
    - Webhook endpoints accept **optional** Bearer tokens for user attribution.
    - Auth service endpoints use **session cookies** (HttpOnly) set after Google OAuth login, or **Bearer JWT tokens** in the Authorization header.
    - Webhook signature headers (e.g. `X-Figma-Signature`, `X-Hub-Signature-256`) are verified when present but not required.
  contact:
    name: RTB AI Hub
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Webhook Listener (local development)
  - url: http://localhost:4001
    description: Auth Service (local development)
...
paths:
...
  /webhooks/figma:
    post:
      tags:
        - Webhook Listener — Figma
      summary: Receive Figma webhook event
      description: |
        Accepts Figma file update events. The pipeline is:
        signature verification (optional) → Bearer auth (optional) → Zod validation → BullMQ queue.
      operationId: receiveFigmaWebhook
      parameters:
        - $ref: '#/components/parameters/FigmaSignature'
        - $ref: '#/components/parameters/OptionalAuthorization'
        - $ref: '#/components/parameters/EnvQuery'
        - $ref: '#/components/parameters/XEnvHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FigmaWebhookPayload'
      responses:
        '202':
          description: Event accepted and queued for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAcceptedResponse'
        '401':
          $ref: '#/components/responses/InvalidSignature'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /webhooks/jira:
    post:
      tags:
        - Webhook Listener — Jira
      summary: Receive Jira webhook event
      description: |
        Accepts Jira issue events (e.g. issue_updated, issue_created). The pipeline is:
        signature verification (optional) → Bearer auth (optional) → Zod validation → BullMQ queue.
      operationId: receiveJiraWebhook
      parameters:
        - $ref: '#/components/parameters/JiraSignature'
        - $ref: '#/components/parameters/OptionalAuthorization'
        - $ref: '#/components/parameters/EnvQuery'
        - $ref: '#/components/parameters/XEnvHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JiraWebhookPayload'
      responses:
        '202':
          description: Event accepted and queued for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAcceptedResponse'
        '401':
          $ref: '#/components/responses/InvalidSignature'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /webhooks/github:
    post:
      tags:
        - Webhook Listener — GitHub
      summary: Receive GitHub webhook event
      description: |
        Accepts GitHub events (pull_request, deployment, etc.). The pipeline is:
        signature verification (optional) → Bearer auth (optional) → Zod validation → BullMQ queue.
      operationId: receiveGitHubWebhook
      parameters:
        - $ref: '#/components/parameters/GitHubSignature'
        - $ref: '#/components/parameters/GitHubEvent'
        - $ref: '#/components/parameters/OptionalAuthorization'
        - $ref: '#/components/parameters/EnvQuery'
        - $ref: '#/components/parameters/XEnvHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GitHubWebhookPayload'
      responses:
        '202':
          description: Event accepted and queued for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAcceptedResponse'
        '401':
          $ref: '#/components/responses/InvalidSignature'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /webhooks/datadog:
    post:
      tags:
        - Webhook Listener — Datadog
      summary: Receive Datadog webhook event
      description: |
        Accepts Datadog alert/event notifications. The pipeline is:
        signature verification (optional) → Bearer auth (optional) → Zod validation → BullMQ queue.
      operationId: receiveDatadogWebhook
      parameters:
        - $ref: '#/components/parameters/DatadogSignature'
        - $ref: '#/components/parameters/OptionalAuthorization'
        - $ref: '#/components/parameters/EnvQuery'
        - $ref: '#/components/parameters/XEnvHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatadogWebhookPayload'
      responses:
        '202':
          description: Event accepted and queued for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAcceptedResponse'
        '401':
          $ref: '#/components/responses/InvalidSignature'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Auth Service (:4001)
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained after Google OAuth login. Used for auth-service endpoints and optionally for webhook attribution.
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: HttpOnly session cookie set after Google OAuth callback.
  parameters:
    FigmaSignature:
      name: X-Figma-Signature
      in: header
      required: false
      description: HMAC signature for Figma webhook payload verification
      schema:
        type: string
    JiraSignature:
      name: X-Hub-Signature
      in: header
      required: false
      description: HMAC signature for Jira webhook payload verification
      schema:
        type: string
    GitHubSignature:
      name: X-Hub-Signature-256
      in: header
      required: false
      description: HMAC-SHA256 signature for GitHub webhook payload verification
      schema:
        type: string
    GitHubEvent:
      name: X-GitHub-Event
      in: header
      required: true
      description: 'GitHub event type (e.g. pull_request, deployment, push)'
      schema:
        type: string
        example: pull_request
    DatadogSignature:
      name: X-Datadog-Signature
      in: header
      required: false
      description: HMAC signature for Datadog webhook payload verification
      schema:
        type: string
    OptionalAuthorization:
      name: Authorization
      in: header
      required: false
      description: 'Optional Bearer JWT token for user attribution (format: `Bearer <token>`)'
      schema:
        type: string
        pattern: '^Bearer .+$'
        example: 'Bearer eyJhbGciOiJIUzI1NiIs...'
    EnvQuery:
      name: env
      in: query
      required: false
      description: 'Target environment for the webhook event. Determines which MCP server containers process the event. Defaults to "int" if not specified.'
      schema:
        type: string
        enum: [int, stg, prd]
        default: int
        example: stg
    XEnvHeader:
      name: X-Env
      in: header
      required: false
      description: 'Alternative way to specify target environment (same as ?env= query parameter). Query parameter takes precedence if both are provided.'
      schema:
        type: string
        enum: [int, stg, prd]
        example: stg
  schemas:
...
    FigmaWebhookPayload:
      type: object
      required:
        - file_key
        - file_name
      properties:
        event_type:
          type: string
          enum: [FILE_UPDATE, FILE_VERSION_UPDATE, FILE_COMMENT]
          default: FILE_UPDATE
          description: Type of Figma event
        file_key:
          type: string
          minLength: 1
          description: Figma file key identifier
          example: abc123def456
        file_name:
          type: string
          minLength: 1
          description: Name of the Figma file
          example: Design System v2
        file_url:
          type: string
          format: uri
          description: Optional direct URL to the Figma file
          example: https://www.figma.com/file/abc123def456
    JiraWebhookPayload:
      type: object
      properties:
        webhookEvent:
          type: string
          default: issue_updated
          description: Jira webhook event name
          example: issue_updated
        issue:
          type: object
          properties:
            key:
              type: string
              minLength: 1
              description: Jira issue key
              example: PROJ-123
            fields:
              type: object
              properties:
                issuetype:
                  type: object
                  properties:
                    name:
                      type: string
                      example: Story
                status:
                  type: object
                  properties:
                    name:
                      type: string
                      example: In Progress
                summary:
                  type: string
                  example: Implement login page
                description:
                  type: string
                  nullable: true
                  example: As a user, I want to log in with my Google account
              additionalProperties: true
        changelog:
          description: Jira changelog object (when fields change)
          nullable: true
      additionalProperties: true
    GitHubWebhookPayload:
      type: object
      properties:
        action:
          type: string
          description: The action that was performed
          example: opened
        repository:
          type: object
          properties:
            full_name:
              type: string
              description: Full repository name (owner/repo)
              example: rtb-team/ai-hub
        pull_request:
          type: object
          properties:
            number:
              type: integer
              description: Pull request number
              example: 42
            title:
              type: string
              description: Pull request title
              example: 'feat: add authentication module'
            head:
              type: object
              properties:
                ref:
                  type: string
                  description: Branch name
                  example: feature/auth
                sha:
                  type: string
                  description: Head commit SHA
                  example: a1b2c3d4e5f6
        deployment:
          type: object
          properties:
            ref:
              type: string
              description: Deployment ref (branch or tag)
            sha:
              type: string
              description: Deployment commit SHA
      additionalProperties: true
    DatadogWebhookPayload:
      type: object
      required:
        - title
      properties:
        event_type:
          type: string
          default: alert
          description: Event type identifier
          example: alert
        id:
          type: string
          description: Datadog event ID
        alert_id:
          type: string
          description: Datadog alert ID
        title:
          type: string
          minLength: 1
          description: Alert or event title
          example: High error rate detected on api-gateway
        body:
          type: string
          description: Event body text
        message:
          type: string
          description: Alert message (alternative to body)
        priority:
          type: string
          enum: [P1, P2, P3]
          default: P2
          description: Alert priority level
        tags:
          type: array
          items:
            type: string
          description: Datadog tags
          example: ['service:api-gateway', 'env:production']
      additionalProperties: true

    # ── Webhook Responses ───────────────────────
    WebhookAcceptedResponse:
      type: object
      required:
        - status
        - eventId
      properties:
        status:
          type: string
          enum: [accepted]
          example: accepted
        eventId:
          type: string
          description: Identifier for the queued event (file_key, issue key, SHA, or alert ID depending on source)
          example: abc123
        env:
          type: string
          enum: [int, stg, prd]
          description: Target environment for this event
          example: int

    # ── Auth Service Schemas ────────────────────
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Internal server error
    ValidationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
                description: JSON path of the invalid field
                example: file_key
              message:
                type: string
                description: Validation error description
                example: 'file_key is required'
    RateLimitErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Too many requests
  responses:
    InvalidSignature:
      description: Webhook signature verification failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Invalid webhook signature
    ValidationError:
      description: Request body validation failed (Zod schema)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            error: Validation failed
            details:
              - path: file_key
                message: 'file_key is required'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitErrorResponse'
          example:
            error: Too many requests
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error